+++ src/user.clj
@@ -1,5 +1,6 @@
 (ns user
-  (:require [hashp.core]))
+  (:require [hashp.core]
+            [clojure.spec.alpha :as s]))
 
 (defn stop-system! [])
 (defn start-system! [])